Changed video buffer from byte array to int array to match main memory.  This gave a minor boost to performance for 8-bit video and a good boost to 32-bit video performance
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@466 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/gui/Render.java
@@ -52,7 +52,7 @@ public class Render {
             public double ratio;
             public float fps;
             public /*Bitu*/int outPitch;
-            public /*Bit8u*/byte[] outWrite;
+            public /*Bit8u*/int[] outWrite;
             public int outWriteOff;
         }
         public SRC src = new SRC();
@@ -142,7 +142,7 @@ public class Render {
             Check_Palette();
         }
 
-        byte[][] outWrite = new byte[1][];
+        int[][] outWrite = new int[1][];
         IntRef outPitch = new IntRef(render.src.outPitch);
         Main.GFX_StartUpdate(outWrite, outPitch);
         render.src.outPitch = outPitch.value;
